Sort out log encoding
Keep the log output in a consistent output format by forcing logging to
use UTF-8.  Forcing stderr to accomplish this wasn't straightforward, so
a new streamHandler was created to create the effect we want.  And an
appropriate test case to verify.

Based on Tim Waugh's original work.

I tweaked the test case to handle the different names for 'UTF8' and
added a check to not encode byte data for writer().