group_manifests: create plugin for group manifests
Add a group_manifests plugin. For now, it fails if asked to create
a group of manifests.

If asked to create a single manifest, it pulls the x86_64 image manifest
from a repository passed as an argument. It then re-uploads the image
manifest to a new location for each image in the tag_conf. It returns the
image manifest.

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>